pyproject.toml
[tool.poetry]
name = "wrike-todoist"
version = "0.1.0"
description = ""
authors = ["Piotr Kilczuk <piotr@kilczuk.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
PyYAML = "^6.0"
google-api-python-client = "^2.164.0"
pendulum = "^3.0.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
wrike-todoist = "wrike_todoist.console:main"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
